refactor(tts/cosyvoice3): fold Qwen2ByteEncoder into the BPE tokenizer
The 56-line GPT-2 byte→unicode shim was its own public enum used only by
Qwen2BpeTokenizer.encode(). Folded it into the tokenizer as a fileprivate
nested enum (`ByteEncoder`). Dropped the unused `unicodeToByte` inverse
table — nothing in the codebase referenced it.

While here, fixed the misleading doc comment on CosyVoice3FrontendFixture:
it claimed to be a "Python-produced parity fixture" but is actually the
generic synthesizer-input DTO that both Phase 1 (load from .safetensors)
and Phase 2 (build in-memory from text + prompt assets) construct before
calling CosyVoice3Synthesizer.synthesize(fixture:).

Documentation/TTS/CosyVoice3.md updated: Qwen2ByteEncoder.swift row
removed; the BpeTokenizer row notes the fileprivate nested type.

Pipeline/Preprocess/ goes from 8 → 7 files. CosyVoice3 tests still
green (16/16 passing).
